+++ pesummary/cli/summarypages.py
@@ -115,7 +115,8 @@ class _GWWebpageGeneration(GWPostProcessing):
             existing_metafile=self.existing,
             add_to_existing=self.add_to_existing,
             existing_file_kwargs=self.existing_file_kwargs,
-            result_files=self.result_files, notes=self.notes
+            result_files=self.result_files, notes=self.notes,
+            pastro_probs=self.pastro_probs
         )
 
     def generate_webpages(self):