+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Bar chart to show the classification probabilities and added them to the
   webpages
 - Generate both source and extrinsic corner plots for gw result files
+- Compute `HasNS` and `HasRemnant` probabilities with the `p_astro` package
 
 ## 0.2.0 [24/10/2019]